+++ CHANGELOG.md
@@ -13,8 +13,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 - Fix `linear` error when evaluating a date property in a feature
+- Fix `transparent` color displayed as black
+- Fix `symbol` color when setting symbol with blendTo after initializing viz
 - Fix `viewportQuantiles` not getting proper buckets
-- Fix `transparent` color
 - Fix `viewportFeatures` error when line vertices were outside the viewport but there was an intersection.
 
 ## [1.0.0] - 2018-11-08